functions/requirements.txt
azure-functions==1.17.0
azure-data-tables
